refactor getInternalConfig to separate file
- can be used for both keyboardAction and pointerAction
