+++ src/pointerAction.ts
@@ -1,3 +1,4 @@
+import {getInternalConfig} from "./config";
 import {
     decrementActiveDropZoneCount,
     incrementActiveDropZoneCount,
@@ -37,12 +38,8 @@ import {getBoundingRectNoTransforms} from "./helpers/intersection";
 import {Item, Options} from "./types";
 import {DraggedEnteredEvent, DraggedLeftEvent, DraggedOverIndexEvent, InternalConfig, Point} from "./internalTypes";
 
-const DEFAULT_DROP_ZONE_TYPE = "--any--";
 const MIN_OBSERVATION_INTERVAL_MS = 100;
 const MIN_MOVEMENT_BEFORE_DRAG_START_PX = 3;
-const DEFAULT_DROP_TARGET_STYLE = {
-    outline: "rgba(255, 255, 102, 0.7) solid 2px"
-};
 
 let originalDragTarget: HTMLElement | undefined;
 let draggedEl: HTMLElement | undefined;
@@ -467,22 +464,6 @@ function cleanupPostDrop() {
     finalizingPreviousDrag = false;
     unlockOriginDzMinDimensions = undefined;
     isDraggedOutsideOfAnyDz = false;
-}
-
-function getInternalConfig(options: Options): InternalConfig {
-    return {
-        items: [...options.items],
-        flipDurationMs: options.flipDurationMs ?? 0,
-        dropAnimationDurationMs: options.flipDurationMs ?? 0,
-        type: options.type ?? DEFAULT_DROP_ZONE_TYPE,
-        dragDisabled: options.dragDisabled ?? false,
-        morphDisabled: options.morphDisabled ?? false,
-        dropFromOthersDisabled: options.dropFromOthersDisabled ?? false,
-        dropTargetStyle: options.dropTargetStyle ?? DEFAULT_DROP_TARGET_STYLE,
-        dropTargetClasses: options.dropTargetClasses ?? [],
-        transformDraggedElement: options.transformDraggedElement ?? (() => {}),
-        centreDraggedOnCursor: options.centreDraggedOnCursor ?? false
-    };
 }
 
 export function dndzone(node: HTMLElement, options: Options) {
